requirements.txt
matplotlib==3.8.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
osqp==0.6.5
protobuf==5.28.2
PyYAML==6.0.2
scipy==1.14.1